# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CastMemberInput": {
        "properties": {
          "base_stage_target_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Stage Target Id"
          },
          "character_id": {
            "title": "Character Id",
            "type": "integer"
          },
          "prominence": {
            "default": 1,
            "minimum": 1.0,
            "title": "Prominence",
            "type": "integer"
          },
          "role": {
            "default": "",
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "character_id"
        ],
        "title": "CastMemberInput",
        "type": "object"
      },
...
      "Generation": {
        "properties": {
          "actual_cost_usd_cents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Cost Usd Cents"
          },
          "aspect_ratio": {
            "title": "Aspect Ratio",
            "type": "string"
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/GenerationAttachment"
            },
            "title": "Attachments",
            "type": "array"
          },
          "base_stage_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Stage Id"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/Candidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "cost_usd_cents": {
            "title": "Cost Usd Cents",
            "type": "integer"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "error_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Text"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "image_size": {
            "title": "Image Size",
            "type": "string"
          },
          "interaction_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interaction Id"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "prompt_hash": {
            "title": "Prompt Hash",
            "type": "string"
          },
          "reserved_cost_usd_cents": {
            "title": "Reserved Cost Usd Cents",
            "type": "integer"
          },
          "scene_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scene Id"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "id",
          "model",
          "image_size",
          "aspect_ratio",
          "prompt",
          "prompt_hash",
          "attachments",
          "warnings",
          "cost_usd_cents",
          "reserved_cost_usd_cents",
          "actual_cost_usd_cents",
          "state",
          "interaction_id",
          "error_text",
          "completed_at",
          "created_at",
          "candidates"
        ],
        "title": "Generation",
        "type": "object"
      },
...
      "GenerationCreate": {
        "properties": {
          "expected_prompt_hash": {
            "title": "Expected Prompt Hash",
            "type": "string"
          }
        },
        "required": [
          "expected_prompt_hash"
        ],
        "title": "GenerationCreate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Scene": {
        "properties": {
          "aspect_ratio": {
            "title": "Aspect Ratio",
            "type": "string"
          },
          "base_stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SceneBaseStage"
              },
              {
                "type": "null"
              }
            ]
          },
          "base_stage_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Stage Id"
          },
          "beat_text": {
            "title": "Beat Text",
            "type": "string"
          },
          "camera": {
            "title": "Camera",
            "type": "string"
          },
          "cast": {
            "items": {
              "$ref": "#/components/schemas/CastMember"
            },
            "title": "Cast",
            "type": "array"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "framing": {
            "title": "Framing",
            "type": "string"
          },
          "generation_count": {
            "title": "Generation Count",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "image_size": {
            "title": "Image Size",
            "type": "string"
          },
          "is_editable": {
            "title": "Is Editable",
            "type": "boolean"
          },
          "latest_attempt_preview_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Attempt Preview Url"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "mood": {
            "title": "Mood",
            "type": "string"
          },
          "style_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Style Id"
          }
        },
        "required": [
          "id",
          "beat_text",
          "camera",
          "framing",
          "mood",
          "aspect_ratio",
          "cast",
          "style_id",
          "base_stage_id",
          "base_stage",
          "model",
          "image_size",
          "created_at",
          "is_editable",
          "generation_count",
          "latest_attempt_preview_url"
        ],
        "title": "Scene",
        "type": "object"
      },
...
      "SceneInput": {
        "properties": {
          "aspect_ratio": {
            "default": "3:2",
            "title": "Aspect Ratio",
            "type": "string"
          },
          "base_stage_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Stage Id"
          },
          "beat_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beat Text"
          },
          "camera": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Camera"
          },
          "cast": {
            "items": {
              "$ref": "#/components/schemas/CastMemberInput"
            },
            "title": "Cast",
            "type": "array"
          },
          "framing": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Framing"
          },
          "image_size": {
            "title": "Image Size",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "mood": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mood"
          },
          "style_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Style Id"
          }
        },
        "required": [
          "cast",
          "model",
          "image_size"
        ],
        "title": "SceneInput",
        "type": "object"
      },
      "SceneModelInput": {
        "properties": {
          "model": {
            "title": "Model",
            "type": "string"
          }
        },
        "required": [
          "model"
        ],
        "title": "SceneModelInput",
        "type": "object"
      },
...
      "SceneSummary": {
        "description": "Lightweight scene reference used where a full scene payload is overkill,\ne.g. \"which scenes use this Base Stage\".",
        "properties": {
          "beat_text": {
            "title": "Beat Text",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "generation_count": {
            "title": "Generation Count",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_editable": {
            "title": "Is Editable",
            "type": "boolean"
          },
          "latest_attempt_preview_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Attempt Preview Url"
          }
        },
        "required": [
          "id",
          "beat_text",
          "created_at",
          "is_editable",
          "generation_count",
          "latest_attempt_preview_url"
        ],
        "title": "SceneSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "LoreCraft3000",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/base-stages/{base_stage_id}/content": {
      "get": {
        "operationId": "base_stage_content_api_v1_base_stages__base_stage_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "base_stage_id",
            "required": true,
            "schema": {
              "title": "Base Stage Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Base Stage Content",
        "tags": [
          "api-v1",
          "api-v1-base-stages"
        ]
      }
    },
...
    "/api/v1/base-stages/{base_stage_id}/scenes": {
      "get": {
        "description": "Scenes whose composition is anchored on this Base Stage.",
        "operationId": "list_base_stage_scenes_api_v1_base_stages__base_stage_id__scenes_get",
        "parameters": [
          {
            "in": "path",
            "name": "base_stage_id",
            "required": true,
            "schema": {
              "title": "Base Stage Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SceneSummary"
                  },
                  "title": "Response List Base Stage Scenes Api V1 Base Stages  Base Stage Id  Scenes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Base Stage Scenes",
        "tags": [
          "api-v1",
          "api-v1-base-stages"
        ]
      }
    },
...
    "/api/v1/scenes": {
      "get": {
        "operationId": "list_scenes_api_v1_scenes_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Scene"
                  },
                  "title": "Response List Scenes Api V1 Scenes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Scenes",
        "tags": [
          "api-v1",
          "api-v1-scenes"
        ]
      },
      "post": {
        "operationId": "create_scene_api_v1_scenes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SceneInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scene"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Scene",
        "tags": [
          "api-v1",
          "api-v1-scenes"
        ]
      }
    },
    "/api/v1/scenes/{scene_id}": {
      "delete": {
        "description": "Permanently delete a scene and its full generation history.",
        "operationId": "delete_scene_api_v1_scenes__scene_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Scene",
        "tags": [
          "api-v1",
          "api-v1-scenes"
        ]
      },
      "get": {
        "operationId": "get_scene_api_v1_scenes__scene_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scene"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scene",
        "tags": [
          "api-v1",
          "api-v1-scenes"
        ]
      },
      "put": {
        "operationId": "update_scene_api_v1_scenes__scene_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SceneInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scene"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Scene",
        "tags": [
          "api-v1",
          "api-v1-scenes"
        ]
      }
    },
    "/api/v1/scenes/{scene_id}/duplicate": {
      "post": {
        "operationId": "duplicate_scene_api_v1_scenes__scene_id__duplicate_post",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scene"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Duplicate Scene",
        "tags": [
          "api-v1",
          "api-v1-scenes"
        ]
      }
    },
    "/api/v1/scenes/{scene_id}/generate": {
      "post": {
        "operationId": "generate_scene_api_v1_scenes__scene_id__generate_post",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Generation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Scene",
        "tags": [
          "api-v1",
          "api-v1-scenes"
        ]
      }
    },
    "/api/v1/scenes/{scene_id}/generations": {
      "get": {
        "operationId": "list_scene_generations_api_v1_scenes__scene_id__generations_get",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GenerationSummary"
                  },
                  "title": "Response List Scene Generations Api V1 Scenes  Scene Id  Generations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Scene Generations",
        "tags": [
          "api-v1",
          "api-v1-scenes"
        ]
      }
    },
    "/api/v1/scenes/{scene_id}/model": {
      "patch": {
        "operationId": "update_scene_model_api_v1_scenes__scene_id__model_patch",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SceneModelInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scene"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Scene Model",
        "tags": [
          "api-v1",
          "api-v1-scenes"
        ]
      }
    },
    "/api/v1/scenes/{scene_id}/preview": {
      "get": {
        "operationId": "preview_scene_api_v1_scenes__scene_id__preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Timezone",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Timezone"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenePreview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Scene",
        "tags": [
          "api-v1",
          "api-v1-scenes"
        ]
      }
    },
...
  }
}